+++ src/hooks/useInfiniteScrollActivity.ts
@@ -3,6 +3,8 @@ import { useCallback, useEffect, useRef, useState } from 'react';
 import { getActivities } from '@/lib/apis/getApis';
 import { MyActivity } from '@/types/get/activityTypes';
 
+import { useInfiniteScrollHandler } from './useInfiniteScrollHandler';
+
 const useInfiniteScrollActivity = () => {
   const [activities, setActivities] = useState<MyActivity[]>([]);
   const [nextCursorId, setNextCursorId] = useState<string | null>(null);
@@ -33,23 +35,15 @@ const useInfiniteScrollActivity = () => {
     [nextCursorId],
   );
 
-  const handleScroll = useCallback(() => {
-    const scrollPosition = window.innerHeight + window.scrollY;
-    const threshold = document.documentElement.offsetHeight - 100;
-
-    if (scrollPosition >= threshold && !loadingRef.current && nextCursorId) {
-      fetchActivitiesData();
-    }
-  }, [fetchActivitiesData, nextCursorId]);
-
   useEffect(() => {
     fetchActivitiesData(true);
   }, []);
 
-  useEffect(() => {
-    window.addEventListener('scroll', handleScroll);
-    return () => window.removeEventListener('scroll', handleScroll);
-  }, [handleScroll]);
+  useInfiniteScrollHandler(
+    fetchActivitiesData,
+    !!nextCursorId,
+    loadingRef.current,
+  );
 
   return {
     activities,
